requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
wikipedia
langchain-text-splitters
pymupdf
unstructured
chromadb
ollama
huggingface
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_Chroma
langchain-openai
beautifulsoup4
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
openai